Extracted Dto's from InventoryServices to new files
