+++ csharp/src/Bargreen.Services/Dtos/AccountingBalance.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Bargreen.Services.Dtos
+{
+
+    public class AccountingBalance
+    {
+        public string ItemNumber { get; set; }
+        public decimal TotalInventoryValue { get; set; }
+    }
+}
