+++ csharp/src/Bargreen.Services/Dtos/InventoryReconciliationResult.cs
@@ -1,0 +1,9 @@
+﻿namespace Bargreen.Services.Dtos
+{
+    public class InventoryReconciliationResult
+    {
+        public string ItemNumber { get; set; }
+        public decimal TotalValueOnHandInInventory { get; set; }
+        public decimal TotalValueInAccountingBalance { get; set; }
+    }
+}
